Polished Small Enemy AI.
Waypoints are dynamically created.

Enemies can now detect edges so that they don't walk off the edge when waypoints are beyond their reach.

If they can't reach waypoint, the enemy will switch to the next waypoint. As of now I'm coding the waypoints system under the assumtion that there are only 2 patrol waypoints.

The implementation for the enemy chasing the player has also been changed